cfs package optimization
